LeafSense_Project/LeafSense_Project_C1SE03_clean/LeafSense_Project/requirement.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==5.0.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.1
cycler==0.12.1
ecdsa==0.19.1
fastapi==0.104.1
filelock==3.19.1
filetype==1.2.0
fonttools==4.60.0
fsspec==2025.9.0
greenlet==3.2.4
h11==0.16.0
idna==3.7
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
mysqlclient==2.2.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
packaging==25.0
passlib==1.7.4
pi_heif==1.1.0
pillow==11.3.0
pillow-avif-plugin==1.5.2
polars==1.33.1
psutil==7.1.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.5.0
pydantic_core==2.14.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
roboflow==1.2.9
rsa==4.9.1
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
ultralytics==8.3.202
ultralytics-thop==2.0.17
urllib3==2.5.0
uvicorn==0.24.0
